bindings: Bindings for ion pair finding function
Binding for James::Path::find_ion_pairs (when the WeightType is double).
Added a test for finding ion pairs and writing them out to a dictionary,
where the keys are the timesteps and the values are the list of ion
pairs for that timestep. The ion pairs can be saved such that the
elements are either atom IDs or indices in the System object (can be
unsafe if the user deletes atoms, reshuffles them etc).