ion_pairs: Also returns n_atoms
The function `get_ion_pairs_time_series` returns a single n_atoms
value, corresponding to the maximum number of atoms in the System object
in the time series, if the number of atoms changes per timestep.
Since the number of atoms can change, it is safer to use the atom IDs.