ion_pairs: List of networks from ion pair info
A list of UndirectedNetwork objects is constructed using the information
in the time series of the ion pairs (of a prescribed length). The end
points of the ion pairs are saved as connections in the network.
`std::span` is not natively supported by pybind11, so I wrote some hacky
wrapper functions for `get_neighbours` and `get_weights` which returned
`std::vector`, which I then wrote bindings for.